session: move btrfs clone main to session and make exit code meaningful
