+++ lib/cfg.h
@@ -151,6 +151,10 @@ typedef struct RmCfg {
 	 * (or directories)
 	 */
 	gboolean run_equal_mode;
+    /* for --btrfs-clone option */
+    bool btrfs_clone;
+    bool btrfs_readonly;
+
 } RmCfg;
 
 /**
