+++ lib/cmdline.c
@@ -233,105 +233,6 @@ static int rm_cmd_maybe_switch_to_hasher(int argc, const char **argv) {
     return EXIT_SUCCESS;
 }
 
-static void rm_cmd_btrfs_clone_usage(void) {
-    rm_log_error(_("Usage: rmlint --btrfs-clone [-r] source dest\n"));
-}
-
-static void rm_cmd_btrfs_clone(const char *source, const char *dest,
-                               const gboolean read_only) {
-#if HAVE_BTRFS_H
-    struct {
-        struct btrfs_ioctl_same_args args;
-        struct btrfs_ioctl_same_extent_info info;
-    } extent_same;
-    memset(&extent_same, 0, sizeof(extent_same));
-
-    int source_fd = rm_sys_open(source, O_RDONLY);
-    if(source_fd < 0) {
-        rm_log_error_line(_("btrfs clone: failed to open source file"));
-        return;
-    }
-
-    extent_same.info.fd = rm_sys_open(dest, read_only ? O_RDONLY : O_RDWR);
-    if(extent_same.info.fd < 0) {
-        rm_log_error_line(_("btrfs clone: error %i: failed to open dest file.%s"),
-                          errno,
-                          read_only ? "" : _("\n\t(if target is a read-only snapshot "
-                                             "then -r option is required)"));
-        rm_sys_close(source_fd);
-        return;
-    }
-
-    struct stat source_stat;
-    fstat(source_fd, &source_stat);
-
-    guint64 bytes_deduped = 0;
-    gint64 bytes_remaining = source_stat.st_size;
-    int ret = 0;
-    while(bytes_deduped < (guint64)source_stat.st_size && ret == 0 &&
-          extent_same.info.status == 0 && bytes_remaining) {
-        extent_same.args.dest_count = 1;
-        extent_same.args.logical_offset = bytes_deduped;
-        extent_same.info.logical_offset = bytes_deduped;
-
-        /* BTRFS_IOC_FILE_EXTENT_SAME has an internal limit at 16MB */
-        extent_same.args.length = MIN(16 * 1024 * 1024, bytes_remaining);
-        if(extent_same.args.length == 0) {
-            extent_same.args.length = bytes_remaining;
-        }
-
-        ret = ioctl(source_fd, BTRFS_IOC_FILE_EXTENT_SAME, &extent_same);
-        if(ret == 0 && extent_same.info.status == 0) {
-            bytes_deduped += extent_same.info.bytes_deduped;
-            bytes_remaining -= extent_same.info.bytes_deduped;
-        }
-    }
-
-    rm_sys_close(source_fd);
-    rm_sys_close(extent_same.info.fd);
-
-    if(ret < 0) {
-        ret = errno;
-        rm_log_error_line(_("BTRFS_IOC_FILE_EXTENT_SAME returned error: (%d) %s"), ret,
-                          strerror(ret));
-    } else if(extent_same.info.status == -22 && read_only && getuid()) {
-        rm_log_error_line(_("Need to run as root user to clone to a read-only snapshot"));
-    } else if(extent_same.info.status < 0) {
-        rm_log_error_line(_("BTRFS_IOC_FILE_EXTENT_SAME returned status %d for file %s"),
-                          extent_same.info.status, dest);
-    } else if(bytes_remaining > 0) {
-        rm_log_info_line(_("Files don't match - not cloned"));
-    }
-#else
-    (void)source;
-    (void)dest;
-    (void)read_only;
-    rm_log_error_line(_("rmlint was not compiled with btrfs support."))
-
-#endif
-}
-
-static int rm_cmd_maybe_btrfs_clone(RmSession *session, int argc, const char **argv) {
-    if(argc > 0 && g_strcmp0("--btrfs-clone", argv[1]) == 0) {
-        /* treat as a btrfs clone subcommand... */
-        if(!rm_session_check_kernel_version(session, 4, 2)) {
-            rm_log_warning_line("This needs at least linux >= 4.2.");
-        } else if(argc == 5 && g_strcmp0("-r", argv[2]) == 0) {
-            /* -r option for deduping read-only snapshots */
-            /* TODO: add check for root user permissions */
-            rm_cmd_btrfs_clone(argv[3], argv[4], TRUE);
-        } else if(argc == 4) {
-            rm_cmd_btrfs_clone(argv[2], argv[3], FALSE);
-        } else {
-            /* malformed command */
-            rm_cmd_btrfs_clone_usage();
-        }
-        /* return EXIT_FAILURE to indicate not to go ahead with main rmlint call */
-        return EXIT_FAILURE;
-    }
-    return EXIT_SUCCESS;
-}
-
 /* clang-format off */
 static const struct FormatSpec {
     const char *id;
@@ -1301,7 +1202,6 @@ static char * rm_cmd_find_own_executable_path(RmSession *session, char **argv) {
 /* Parse the commandline and set arguments in 'settings' (glob. var accordingly) */
 bool rm_cmd_parse_args(int argc, char **argv, RmSession *session) {
     RmCfg *cfg = session->cfg;
-    gboolean clone = FALSE;
 
     /* Handle --gui before all other processing,
      * since we need to pass other args to the python interpreter.
@@ -1313,10 +1213,6 @@ bool rm_cmd_parse_args(int argc, char **argv, RmSession *session) {
     }
 
     if(rm_cmd_maybe_switch_to_hasher(argc, (const char **)argv) == EXIT_FAILURE) {
-        return false;
-    }
-
-    if(rm_cmd_maybe_btrfs_clone(session, argc, (const char **)argv) == EXIT_FAILURE) {
         return false;
     }
 
@@ -1378,14 +1274,15 @@ bool rm_cmd_parse_args(int argc, char **argv, RmSession *session) {
         {"no-hardlinked"            , 'L'  , DISABLE   , G_OPTION_ARG_NONE      , &cfg->find_hardlinked_dupes    , _("Ignore hardlink twins")                                                , NULL}     ,
         {"partial-hidden"           , 0    , EMPTY     , G_OPTION_ARG_CALLBACK  , FUNC(partial_hidden)           , _("Find hidden files in duplicate folders only")                          , NULL}     ,
         {"mtime-window"             , 'Z'  , 0         , G_OPTION_ARG_DOUBLE    , &cfg->mtime_window             , _("Consider duplicates only equal when mtime differs at max. T seconds")  , "T"}      ,
+        {"btrfs-clone"              , 0    , 0         , G_OPTION_ARG_NONE      , &cfg->btrfs_clone              , _("Clone extents from source to dest, if extents match")                  , NULL}     ,
+        {"btrfs-readonly"           , 'r'  , 0         , G_OPTION_ARG_NONE      , &cfg->btrfs_readonly           , _("(btrfs-clone option) also clone to read-only snapshots (needs root)")  , NULL}     ,
 
         /* Callback */
         {"show-man" , 'H' , EMPTY , G_OPTION_ARG_CALLBACK , rm_cmd_show_manpage , _("Show the manpage")            , NULL} ,
         {"version"  , 0   , EMPTY , G_OPTION_ARG_CALLBACK , rm_cmd_show_version , _("Show the version & features") , NULL} ,
         /* Dummy option for --help output only: */
         {"gui"         , 0 , 0 , G_OPTION_ARG_NONE , NULL   , _("If installed, start the optional gui with all following args")                 , NULL},
-        {"hash"        , 0 , 0 , G_OPTION_ARG_NONE , NULL   , _("Work like sha1sum for all supported hash algorithms (see also --hash --help)") , NULL}                                            ,
-        {"btrfs-clone" , 0 , 0 , G_OPTION_ARG_NONE , &clone , _("Clone extents from source to dest, if extents match")                          , NULL} ,
+        {"hash"        , 0 , 0 , G_OPTION_ARG_NONE , NULL   , _("Work like sha1sum for all supported hash algorithms (see also --hash --help)") , NULL},
 
         /* Special case: accumulate leftover args (paths) in &paths */
         {G_OPTION_REMAINING , 0 , 0 , G_OPTION_ARG_FILENAME_ARRAY , &paths , ""   , NULL}   ,
@@ -1497,12 +1394,6 @@ bool rm_cmd_parse_args(int argc, char **argv, RmSession *session) {
 
     if(!g_option_context_parse(option_parser, &argc, &argv, &error)) {
         goto failure;
-    }
-
-    if(clone) {
-        /* should not get here */
-        rm_cmd_btrfs_clone_usage();
-        session->cmdline_parse_error = TRUE;
     }
 
     /* Silent fixes of invalid numeric input */
