+++ lib/session.c
@@ -33,6 +33,11 @@
 #include "session.h"
 #include "traverse.h"
 
+#if HAVE_BTRFS_H
+#include <linux/btrfs.h>
+#include <sys/ioctl.h>
+#endif
+
 #if HAVE_UNAME
 #include "sys/utsname.h"
 
@@ -99,8 +104,8 @@ void rm_session_init(RmSession *session, RmCfg *cfg) {
     session->timer_since_proc_start = g_timer_new();
     g_timer_start(session->timer_since_proc_start);
 
-	/* Assume that files are not equal */
-	session->equal_exit_code = EXIT_FAILURE;
+    /* Assume that files are not equal */
+    session->equal_exit_code = EXIT_FAILURE;
 }
 
 void rm_session_clear(RmSession *session) {
@@ -160,3 +165,109 @@ bool rm_session_was_aborted() {
 
     return rc;
 }
+/**
+ * *********** btrfs clone session main ************
+ **/
+int rm_session_btrfs_clone_main(RmSession *session) {
+    RmCfg *cfg = session->cfg;
+    if(cfg->path_count != 2) {
+        rm_log_error(_("Usage: rmlint --btrfs-clone [-r] [-v|V] source dest\n"));
+        return EXIT_FAILURE;
+    }
+
+    if(!rm_session_check_kernel_version(session, 4, 2)) {
+        rm_log_warning_line("This needs at least linux >= 4.2.");
+        return EXIT_FAILURE;
+    }
+
+    /* TODO: if kernel version >= 4.5 then use IOCTL-FIDEDUPERANGE
+     * http://man7.org/linux/man-pages/man2/ioctl_fideduperange.2.html
+     */
+#if HAVE_BTRFS_H
+
+    g_assert(cfg->paths);
+    RmPath *dest = cfg->paths->data;
+    g_assert(cfg->paths->next);
+    RmPath *source = cfg->paths->next->data;
+    rm_log_debug_line("Cloning %s -> %s", source->path, dest->path);
+
+    struct {
+        struct btrfs_ioctl_same_args args;
+        struct btrfs_ioctl_same_extent_info info;
+    } extent_same;
+    memset(&extent_same, 0, sizeof(extent_same));
+
+    int source_fd = rm_sys_open(source->path, O_RDONLY);
+    if(source_fd < 0) {
+        rm_log_error_line(_("btrfs clone: failed to open source file"));
+        return EXIT_FAILURE;
+    }
+
+    extent_same.info.fd =
+        rm_sys_open(dest->path, cfg->btrfs_readonly ? O_RDONLY : O_RDWR);
+    if(extent_same.info.fd < 0) {
+        rm_log_error_line(
+            _("btrfs clone: error %i: failed to open dest file.%s"),
+            errno,
+            cfg->btrfs_readonly ? "" : _("\n\t(if target is a read-only snapshot "
+                                         "then -r option is required)"));
+        rm_sys_close(source_fd);
+        return EXIT_FAILURE;
+    }
+
+    /* fsync's needed to flush extent mapping */
+    fsync(source_fd);
+    fsync(extent_same.info.fd);
+
+    struct stat source_stat;
+    fstat(source_fd, &source_stat);
+
+    guint64 bytes_deduped = 0;
+    gint64 bytes_remaining = source_stat.st_size;
+    int ret = 0;
+    while(bytes_deduped < (guint64)source_stat.st_size && ret == 0 &&
+          extent_same.info.status == 0 && bytes_remaining) {
+        extent_same.args.dest_count = 1;
+        extent_same.args.logical_offset = bytes_deduped;
+        extent_same.info.logical_offset = bytes_deduped;
+
+        /* try to dedupe the rest of the file */
+        extent_same.args.length = bytes_remaining;
+
+        ret = ioctl(source_fd, BTRFS_IOC_FILE_EXTENT_SAME, &extent_same);
+        bytes_deduped += extent_same.info.bytes_deduped;
+        bytes_remaining -= extent_same.info.bytes_deduped;
+        rm_log_debug_line("deduped %lu bytes...", bytes_deduped);
+    }
+
+    rm_sys_close(source_fd);
+    rm_sys_close(extent_same.info.fd);
+
+    if(ret >= 0 && bytes_remaining == 0) {
+        return EXIT_SUCCESS;
+    }
+
+    if(ret < 0) {
+        ret = errno;
+        rm_log_error_line(_("BTRFS_IOC_FILE_EXTENT_SAME returned error: (%d) %s"), ret,
+                          strerror(ret));
+    } else if(extent_same.info.status == -22 && cfg->btrfs_readonly && getuid()) {
+        rm_log_error_line(_("Need to run as root user to clone to a read-only snapshot"));
+    } else if(extent_same.info.status < 0) {
+        rm_log_error_line(_("BTRFS_IOC_FILE_EXTENT_SAME returned status %d for file %s"),
+                          extent_same.info.status, dest->path);
+    } else if(bytes_deduped == 0) {
+        rm_log_info_line(_("Files don't match - not cloned"));
+    } else if(bytes_remaining > 0) {
+        rm_log_info_line(_("Only first %lu bytes cloned - files not fully identical"),
+                         bytes_deduped);
+    }
+
+#else
+    (void)cfg;
+    rm_log_error_line(_("rmlint was not compiled with btrfs support."))
+#endif
+
+    return EXIT_FAILURE;
+}
+
